UI screens linked, and more work done on UI
An issue with borders on the game screen and an issue with icons on the title screen still needs to be fixed.
And the button on the title screen is still a placeholder for the text resize animator unless we want to go with the button. 
The high score display still has to be implemented, as well as the game board block check/ pause pop-up/game over--> high score display.