Javadoc and fixed nextQueuePanel bugs.
Notes: User must have a copy of userList.ser to save High Scores locally. I recommend adding the userList.ser from the original master branch to get the program to run.
Changes: I changed the queue to an array of Pieces in the board class and changed the NextPieceQueue constructor to match the input.